Placate Swiftlint
