+++ Sources/ResourceProvider/Operators/Concurrent.swift
@@ -20,8 +20,8 @@ public extension SyncProvider where Self: Sendable {
      If a sync provider needs to be used in an `async` context and it doesn't play well with concurrency —usually
      because you want to avoid data races with its state management— you will want to wrap it in one of these.
 
-     - TODO: Talk about required sendability of ID & Value.
-     - TODO: Talk about IKWID for making valueForID functionally `@Sendable`
+     - Todo: Talk about required sendability of ID & Value.
+     - Todo: Talk about IKWID for making valueForID functionally `@Sendable`
      - Returns: An `async` provider version of the calling `SyncProvider` that runs its calls serially.
      */
     func concurrent() -> some AsyncProvider<ID, Value, Failure> {
